refactor: extract tooltip content
